+++ packages/transaction-controller/src/types.ts
@@ -1900,3 +1900,46 @@ export type AssetFiatValues = {
    */
   sending?: string;
 };
+
+/**
+ * Options for adding a transaction.
+ */
+export type AddTransactionOptions = {
+  /** Unique ID to prevent duplicate requests. */
+  actionId?: string;
+  /** The fiat values of the assets in the transaction. */
+  assetsFiatValues?: AssetFiatValues;
+  /** A custom ID for the batch this transaction belongs to. */
+  batchId?: Hex;
+  /** An enum to indicate what device confirmed the transaction. */
+  deviceConfirmedOn?: WalletDevice;
+  /** Whether to disable the gas estimation buffer. */
+  disableGasBuffer?: boolean;
+  /** RPC method that requested the transaction. */
+  method?: string;
+  /** Params for any nested transactions encoded in the data. */
+  nestedTransactions?: NestedTransactionMetadata[];
+  /** The id of the network client for this transaction. */
+  networkClientId: NetworkClientId;
+  /** The origin of the transaction request, such as a dApp hostname. */
+  origin?: string;
+  /** Custom logic to publish the transaction. */
+  publishHook?: PublishHook;
+  /** Whether the transaction requires approval by the user, defaults to true unless explicitly disabled. */
+  requireApproval?: boolean | undefined;
+  /** Response from security validator. */
+  securityAlertResponse?: SecurityAlertResponse;
+  /** The sendFlowHistory entries to add. */
+  sendFlowHistory?: SendFlowHistoryEntry[];
+  /** Options for swaps transactions. */
+  swaps?: {
+    /** Whether the transaction has an approval transaction. */
+    hasApproveTx?: boolean;
+    /** Metadata for swap transaction. */
+    meta?: Partial<TransactionMeta>;
+  };
+  /** The parent context for any new traces. */
+  traceContext?: unknown;
+  /** Type of transaction to add, such as 'cancel' or 'swap'. */
+  type?: TransactionType;
+};